Filters moved to shared
This is prep for making card searches available via the api
