+++ packages/frontend/src/hellfall/TODO.md
@@ -81,10 +81,8 @@ Write a script to transform the database
 - add button to expose card json? and copy/pastable card details?
 - add {69}, {45}, {-2}
 - make card name text render in hellfallentry below card (so if card doesn't render, the name will)
-- add card to plaintext function; add it as alt text to card images (do by face/all faces)
 - add collections to deal with types, tags, keywords, etc.
 - add minimize button next to close button on card pane (turns it into small popup at bottom with only cardname) (also add ability to store multiple like this) (use minus symbol)
-- add button in pane to open card in new tab (use box with arrow)
 - don't trigger new history on activecard change?
 - figure out if pip shadows have correct angle (should they be slightly more downward than leftward?)
 - implement flip/aftermath rotation buttons (base it off of cockatrice face merge logic)
