+++ packages/frontend/src/hellfall/search-controls/SortComponent.tsx
@@ -4,15 +4,15 @@ import { SecondaryButton, TertiaryButton, PrimaryButton } from '@workday/canvas-
 import { useAtom, useAtomValue } from 'jotai';
 import { inputSortAtom, queryAtom, querySortAtom, sortAtom } from '../atoms/searchAtoms.ts';
 import styled from '@emotion/styled';
-import { sorts, dirs, sortType, dirType } from '../filters/types.ts';
+import { sorts, dirs, sortType, dirType } from '@hellfall/shared/filters/types.ts';
 import {
   getWinnowedSortOptions,
   combineAndWinnowSorts,
   parseSorts,
-} from '../filters/parseSearchBar.ts';
+} from '@hellfall/shared/filters/parseSearchBar.ts';
 import { useEffect, useState } from 'react';
 import { plusIcon, minusIcon } from '@workday/canvas-system-icons-web';
-import { makeSort } from '../filters/filterBuilder.ts';
+import { makeSort } from '@hellfall/shared/filters/filterBuilder.ts';
 
 const ALL_SORT_OPTIONS: Array<{ label: string; value: sortType }> = [
   { label: 'Auto', value: 'auto' },
