+++ packages/shared/package.json
@@ -21,6 +21,10 @@
     "./utils": {
       "types": "./src/utils/index.ts",
       "import": "./src/utils/index.ts"
+    },
+    "./filters": {
+      "types": "./src/filters/index.ts",
+      "import": "./src/filters/index.ts"
     }
   }
 }
