+++ packages/shared/src/utils/cardHandling.ts
@@ -1,4 +1,5 @@
 import { HCCard, HCCardFace, HCColors } from '../types';
+import { listShareLower } from './listHandling';
 import { facePropOrder, partPropOrder, propOrder } from './orderProps';
 import { getIndicatorFromColors } from './pipsHandling';
 import {
@@ -16,6 +17,7 @@ import {
   propType,
   valueType,
 } from './propTypes';
+import { getMasterpiece, getSetCode, stripMasterpiece, stripSetCode } from './textHandling';
 
 /**
  * Converts the card to an array of its faces.
@@ -182,3 +184,71 @@ export const toPlainText = (card: HCCard.Any) =>
   toFaces(card)
     .map(face => faceToPlainText(face))
     .join('\n---\n');
+
+export const getOtherNames = (card: HCCard.Any): string[] | undefined => {
+  const names = [];
+  if (card.tags?.includes('irregular-face-name') && 'card_faces' in card) {
+    names.push(card.card_faces.map(face => face.name).join(' \\ '));
+  }
+  if (card.flavor_name) {
+    names.push(card.flavor_name);
+  }
+  if ('card_faces' in card && card.card_faces.find(face => face.flavor_name)) {
+    names.push(
+      ...(card.card_faces
+        .filter(face => face.flavor_name)
+        .map(face => face.flavor_name) as string[])
+    );
+  }
+  return names.length ? names : undefined;
+};
+
+export const getAllNames = (card: HCCard.Any): string[] => {
+  const names = [stripMasterpiece(stripSetCode(card.name))];
+  if (names[0].slice(-5) == ' (HC)') {
+    names.push(card.name);
+    while (names[0].slice(-5) == ' (HC)') {
+      names.unshift(stripSetCode(names[0]));
+    }
+  }
+  if (card.flavor_name) {
+    names.push(card.flavor_name);
+  }
+  const start = getMasterpiece(card.name);
+  const ending = getSetCode(card.name);
+  if ('card_faces' in card) {
+    card.card_faces.forEach((face, i) => {
+      if (face.flavor_name) {
+        names.push(face.flavor_name);
+      }
+      names.push(face.name);
+      let name = face.name;
+      card.card_faces.slice(i + 1).forEach(f => {
+        name += ` // ${f.name}`;
+        names.push(name);
+      });
+    });
+  }
+  if (start || ending) {
+    return names.flatMap(name => {
+      return [
+        name,
+        ...(start ? [start + name] : []),
+        ...(ending ? [name + ending] : []),
+        ...(start && ending ? [start + name + ending] : []),
+      ];
+    });
+  }
+  return names;
+};
+
+export const canBeACommander = (card: HCCard.Any) => {
+  const faces = toFaces(card);
+  return (
+    ((listShareLower(faces[0].supertypes, 'legendary') &&
+      (listShareLower(faces[0].types, 'creature') ||
+        listShareLower(faces[0].subtypes, ['vehicle', 'spacecraft', 'watercraft']))) ||
+      faces[0]?.oracle_text.toLowerCase().includes('can be your commander')) &&
+    !faces[0]?.oracle_text.toLowerCase().includes('irresponsible')
+  );
+};
