+++ packages/shared/src/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './isInt.ts';
+export * from './numHandling.ts';
 export * from './listHandling.ts';
 export * from './orderColors.ts';
 export * from './textHandling.ts';
